REFACTOR UNIT TEST - Rename the file name and namespace
